+++ scripts/macros/actions/nonLethal.js
@@ -1,4 +1,4 @@
-import {chris} from '../../../helperFunctions.js';
+import {chris} from '../../helperFunctions.js';
 export async function nonLethal({speaker, actor, token, character, item, args, scope, workflow}) {
     if (workflow.targets.size != 1) return;
     let validTypes = [