+++ scripts/macros/actions/underwater.js
@@ -1,5 +1,5 @@
-import {chris} from '../../../helperFunctions.js';
-import {queue} from '../../../utility/queue.js';
+import {chris} from '../../helperFunctions.js';
+import {queue} from '../../utility/queue.js';
 async function attack({speaker, actor, token, character, item, args, scope, workflow}) {
     switch (workflow.item.system.actionType) {
         case 'mwak':